Fix BodySlide slider parsing: read fractional values as float
SynthEBD read <SetSlider value="..."> with int.TryParse, which returns
false for any value containing a decimal point. BodySlide writes those
routinely, so the whole SetSlider was skipped and the slider silently
stayed at 0, with no error surfaced anywhere. 1000 of the 5094 presets
in the measurement cache (19.6%) lost at least one non-zero slider; the
worst kept 4 of 320.

There is one parse site and one deform site, and the measurement
evaluator reads geometry off the renderer's deformed mesh, so this
mangled the 3D preview and every Body Type Profile measurement together.

BodySlideSlider.Big/Small are now float, parsed through a new
Settings_OBody.TryParseSliderValue. It pins the invariant culture --
values always use '.' separators, so current-culture parsing would
reintroduce the same silent drop on comma-decimal locales -- and it
rejects non-finite input, since float.TryParse accepts "NaN"/"Infinity"
where int.TryParse did not and a NaN slider poisons every vertex that
slider touches.

Consumers: BodyGenSpecsParser drops a now-pointless round-to-int;
MeasurementCacheStore formats the fingerprint round-trip/invariant so it
never depends on locale; the annotator preview shows float endpoints;
the legacy trainer export rounds at the boundary to keep its int[]
schema.

Out-of-range values (200, -160) are still applied literally, matching
the deformer and confirmed against real BodySlide output.

Verified against the live measurement cache: recomputing the preset
fingerprint under both the old and new rules re-hashes exactly the 1000
mis-parsed presets and no others. After regenerating the cache, all
movement is confined to those presets and the tuned classifier
thresholds are unchanged.

+++ SynthEBD/Classes_Core/ViewModels/OBody SubModels/VM_SliderAnnotatorPreviewPanel.cs
@@ -539,10 +539,10 @@ public class VM_AnnotatorPresetRow : VM
     public string Label => PlaceHolder.Label ?? "";
 
     /// <summary>Chosen slider's authored weight-0 value; null when the preset lacks the slider.</summary>
-    public int? Low { get; private set; }
+    public float? Low { get; private set; }
 
     /// <summary>Chosen slider's authored weight-100 value; null when the preset lacks the slider.</summary>
-    public int? High { get; private set; }
+    public float? High { get; private set; }
 
     /// <summary>Chosen slider's linear blend at the panel's preview weight; null when the preset lacks the slider.</summary>
     public float? Interpolated { get; private set; }
@@ -584,8 +584,8 @@ public class VM_AnnotatorSliderValueRow : VM
     }
 
     public string SliderName { get; }
-    public int Low { get; }
-    public int High { get; }
+    public float Low { get; }
+    public float High { get; }
     public float Interpolated { get; private set; }
 
     /// <summary>Recomputes <see cref="Interpolated"/> when the panel's preview weight changes.</summary>
